Update README.md
Dynamic English Dictionary is a modern and intuitive web application that enables users to quickly search for English word definitions, view phonetics, hear pronunciation audio, and explore example sentences. Designed with responsiveness and clarity in mind, the app leverages Bootstrap and JavaScript to provide a smooth and visually appealing user experience. It connects to the open and free [dictionaryapi.dev](https://dictionaryapi.dev/) to retrieve accurate and up-to-date information, making it an ideal tool for learners, educators, and everyday users alike.